jmr pointed out that i fail
git-svn-id: https://svn.macports.org/repository/macports/trunk/base@49220 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ src/tclobjc1.0/tclobjc.m
@@ -115,7 +115,7 @@ int tclobjc_dispatch(ClientData clientData, Tcl_Interp *interp, int objc, Tcl_Ob
 		} else if (arg_type[0] == _C_UINT) {
             long value;
             if (Tcl_GetLongFromObj(interp, objv[i], &value) == TCL_OK) {
-                if ((unsigned int)value > UINT_MAX || value < 0) {
+                if ((unsigned long)value > UINT_MAX || value < 0) {
                     NSString *str = [NSString stringWithFormat:@"Unsigned integer argument invalid: %ld", value];
                     Tcl_Obj *tcl_result = Tcl_NewStringObj([str UTF8String], -1);
                     Tcl_SetObjResult(interp, tcl_result);
